+++ demo/app.controller.ts
@@ -1,10 +1,14 @@
 import { Controller, Get, headers, inject } from '../mod.ts';
 
+import { AppConfig } from './app.config.ts';
 import { SharedService } from './shared/shared.service.ts';
 
 @Controller()
 export class AppController {
-  constructor(private readonly _sharedService = inject(SharedService)) {}
+  constructor(
+    private readonly _sharedService = inject(SharedService),
+    private readonly _config = inject(AppConfig),
+  ) {}
 
   @Get([headers<string>('user-agent')])
   get(userAgent: string) {
@@ -14,5 +18,10 @@ export class AppController {
   @Get('shared')
   getShared() {
     return this._sharedService.content;
+  }
+
+  @Get('config')
+  getConfig() {
+    return this._config.value;
   }
 }